Move to incorporate changes into existing SDMX-JSON model
