+++ lib/Models/SdmxJsonCatalogItem.js
@@ -13,8 +13,9 @@ var defineProperties = require("terriajs-cesium/Source/Core/defineProperties");
 var deprecationWarning = require("terriajs-cesium/Source/Core/deprecationWarning");
 var freezeObject = require("terriajs-cesium/Source/Core/freezeObject");
 var knockout = require("terriajs-cesium/Source/ThirdParty/knockout");
-var Resource = require("terriajs-cesium/Source/Core/resource");
 var loadJson = require("../Core/loadJson");
+var loadSdmxStructureJson = require("../Core/loadSdmxStructureJson");
+var loadSdmxDataJson = require("../Core/loadSdmxDataJson");
 var when = require("terriajs-cesium/Source/ThirdParty/when");
 
 var arrayProduct = require("../Core/arrayProduct");
@@ -91,6 +92,13 @@ var SdmxJsonCatalogItem = function(terria, url) {
    * @type {String}
    */
   this.dataUrlComponent = undefined;
+
+  /**
+   * Gets or sets the SDMX version number
+   * Defaults to 2.0.
+   * @type {Number}
+   */
+  this.sdmxVersionNumber = undefined;
 
   /**
    * Gets or sets the 'dataflow' SDMX URL component, eg. 'dataflow' in http://stats.oecd.org/sdmx-json/dataflow/QNA.
@@ -516,12 +524,41 @@ function getDataflowUrl(item) {
     // If it contains a /, and anything after the /, then treat it as #1.
     if (suffix.indexOf("/") >= 0 && suffix.indexOf("/") < suffix.length - 1) {
       return;
+    } else if (item.sdmxVersionNumber !== 2.0) {
+      return `${item.restEndpointRoot}${item.dataflowUrlComponent}/${
+        item.agencyId
+      }/${item.flowId}/${
+        item.dataflowVersion
+      }/?references=all&detail=referencepartial`;
     } else {
       // return the same URL but with /data/ replaced with /dataflow/.
       var dataflowUrlComponent = "/" + item.dataflowUrlComponent + "/";
       return item._originalUrl.replace(dataUrlComponent, dataflowUrlComponent);
     }
   }
+}
+
+// https://github.com/sdmx-twg/sdmx-rest/blob/master/v2_1/ws/rest/docs/4_4_data_queries.md#parameters-used-for-identifying-a-resource
+function parseResourceComponents(item) {
+  const uri = new URI(item.url);
+  const segments = uri.segment();
+
+  const flowRef = segments[segments.length - 1].split(",");
+  if (flowRef.length === 1) {
+    item.flowId = flowRef[0];
+  } else if (flowRef.length === 1) {
+    item.agencyId = flowRef[0];
+    item.flowId = flowRef[1];
+  } else {
+    item.agencyId = flowRef[0];
+    item.flowId = flowRef[1];
+    item.dataflowVersion = flowRef[2];
+  }
+}
+
+// return something like https://stats.spc.int/SeptemberDisseminateNSIService/Rest/
+function parseEndpointRoot(item) {
+  return item.url.split("Rest")[0].concat("Rest/");
 }
 
 /*
@@ -563,8 +600,24 @@ function getDataflowUrl(item) {
  *   And result.dataSets[0].series[key].observations[t] has lots of values for different t, not necessarily including t = 0. (eg. key = "21:0:0:0" starts at t = 140).
  */
 SdmxJsonCatalogItem.prototype._load = function() {
+  parseResourceComponents(this);
+
+  this.restEndpointRoot = defaultValue(
+    this.restEndpointRoot,
+    parseEndpointRoot(this)
+  );
+
+  this.agencyId = defaultValue(this.agencyId, "all");
+  this.dataflowVersion = defaultValue(this.dataflowVersion, "latest");
+
   // Set some defaults.
   this._originalUrl = this.originalUrl; // Since `this.url` is often set after initialization.
+
+  this.sdmxVersionNumber = defaultValue(this.sdmxVersionNumber, 2.0);
+
+  this._getData = this.sdmxVersionNumber === 2.1 ? loadSdmxDataJson : loadJson;
+  this._getMetadata =
+    this.sdmxVersionNumber === 2.1 ? loadSdmxStructureJson : loadJson;
   this.regionTypeDimensionId = defaultValue(
     this.regionTypeDimensionId,
     "REGIONTYPE"
@@ -737,6 +790,82 @@ function buildDimensions(item, structureSeries) {
     result[keyPosition] = {
       id: thisSeries.id,
       name: thisSeries.name,
+      // Eg. values: [{id: "BD_2", name: "Births"}, {id: "BD_4", name: "Deaths"}].
+      values: values
+    };
+  }
+  return result;
+}
+
+function getCodeListById(name, codelists) {
+  for (var i = 0; i < codelists.length; i++) {
+    if (codelists[i].id === "CL_" + name) return codelists[i];
+  }
+  return null;
+}
+
+function buildDimensionsFor21(item, structureSeries) {
+  // getFilter returns a function which can be used in list.filter().
+  // It tests if the value is in the given list, using regexps if possible.
+  // filterList can be either item.whitelist or item.blacklist.
+  // set isWhiteList false if is blacklist, so that the return values are negated (except for a missing list).
+  function getFilter(filterList, dimensionId, isWhiteList) {
+    var thisIdsFilterList = filterList[dimensionId];
+    if (!defined(thisIdsFilterList)) {
+      return function() {
+        return true;
+      };
+    }
+    try {
+      var thisIdsRegExps = thisIdsFilterList.map(
+        string => new RegExp("^" + string + "$")
+      );
+      return function(value) {
+        // Test as a straight string, and if that fails, as a regular expression.
+        var isPresent =
+          thisIdsFilterList.indexOf(value.id) >= 0 ||
+          thisIdsRegExps.map(regExp => value.id.match(regExp)).some(defined);
+        if (isWhiteList) {
+          return isPresent;
+        }
+        return !isPresent;
+      };
+    } catch (e) {
+      // Cannot intepret as a regular expression.
+      // Eg. "[" causes Uncaught SyntaxError: Invalid regular expression: /[/: Unterminated character class(…)),
+      // So just test as a string.
+      return function(value) {
+        var isPresent = thisIdsFilterList.indexOf(value.id) >= 0;
+        if (isWhiteList) {
+          return isPresent;
+        }
+        return !isPresent;
+      };
+    }
+  }
+  var result = [];
+  for (var i = 0; i < structureSeries.length; i++) {
+    var thisSeries = structureSeries[i];
+    var keyPosition = defined(thisSeries.keyPosition)
+      ? thisSeries.keyPosition
+      : i; // Since time_period can be an observation, without a keyPosition.
+    const cl = getCodeListById(thisSeries.id, item._codeLists);
+    const values =
+      cl === null
+        ? []
+        : cl.codes.map(function(code) {
+            return {
+              id: code.id,
+              name: code.name.en
+            };
+          });
+    moveTotalValueIdsToFront(
+      values,
+      getTotalValueIdsForDimensionId(item, thisSeries.id)
+    );
+    result[keyPosition] = {
+      id: thisSeries.id,
+      name: thisSeries.id,
       // Eg. values: [{id: "BD_2", name: "Births"}, {id: "BD_4", name: "Deaths"}].
       values: values
     };
@@ -1405,6 +1534,9 @@ function calculateDimensionRequestString(item, conceptIds, fullDimensions) {
       // These do not appear as concepts - directly supply the only value's id.
       return [dimension.values[0].id];
     }
+    if (item.sdmxVersionNumber === 2.1 && "SPC_UNIT_MEASURE" === dimension.id) {
+      return ["."];
+    }
     return conceptIds[nextConceptIndex++];
   });
   if (dimensionRequestArrays.some(a => !defined(a))) {
@@ -1463,7 +1595,11 @@ function getUrlFromDimensionRequestString(item, dimensionRequestString) {
   if (url[url.length - 1] !== "/") {
     url += "/";
   }
-  url += dimensionRequestString + "/" + item.providerId;
+  url +=
+    item.sdmxVersionNumber === 2.1
+      ? dimensionRequestString + "/"
+      : dimensionRequestString + "/" + item.providerId;
+
   if (defined(item.startTime)) {
     url += "?startTime=" + item.startTime;
     if (defined(item.endTime)) {
@@ -1472,21 +1608,31 @@ function getUrlFromDimensionRequestString(item, dimensionRequestString) {
   } else if (defined(item.endTime)) {
     url += "?endTime=" + item.endTime;
   }
+  if (item.sdmxVersionNumber === 2.1)
+    url += "dimensionAtObservation=AllDimensions";
   return url;
 }
 
 // This is called when the URL gives a datasetId, but no specifics.
 // We start by loading in the structure (without any data) from the dataflow URL.
 function loadDataflow(item) {
-  var dataflowUrl = cleanAndProxyUrl(item, item._dataflowUrl);
-  return loadJson(createSdmxJsonResoure(dataflowUrl)).then(function(json) {
+  var dataflowUrl = proxyCatalogItemUrl(item, item._dataflowUrl);
+  return item._getMetadata(dataflowUrl).then(function(json) {
+    if (item.sdmxVersionNumber === 2.1) json = json.data;
     // Then access:
     //   - result.structure.dimensions.observation[k] for {keyPosition, id, name, values[]} to get the name & id of dimension keyPosition and its array of allowed values (with {id, name}).
     //   - result.structure.dimensions.attributes.dataSet has some potentially interesting things such as units, unit multipliers, reference periods (eg. http://stats.oecd.org/sdmx-json/dataflow/QNA).
     //   - result.structure.dimensions.attributes.observation has some potentially interesting things such as time formats and status (eg. estimated value, forecast value).
-    var structureSeries = json.structure.dimensions.observation;
-
-    item._fullDimensions = buildDimensions(item, structureSeries);
+    var structureSeries =
+      item.sdmxVersionNumber === 2.1
+        ? json.dataStructures[0].dataStructureComponents.dimensionList
+            .dimensions
+        : json.structure.dimensions.observation;
+    if (item.sdmxVersionNumber === 2.1) item._codeLists = json.codelists;
+    item._fullDimensions =
+      item.sdmxVersionNumber === 2.1
+        ? buildDimensionsFor21(item, structureSeries)
+        : buildDimensions(item, structureSeries);
     if (
       !defined(
         getDimensionIndexById(item._fullDimensions, item.regionDimensionId)
@@ -1536,7 +1682,10 @@ function buildDimensionsFromPreparedJson(item, preparedJson) {
   var structureSeries = structureDimensionsAndAttributes.filter(function(s) {
     return s._type === "dimensions";
   });
-  return buildDimensions(item, structureSeries);
+
+  return item.sdmxVersionNumber === 2.1
+    ? structureSeries
+    : buildDimensions(item, structureSeries);
 }
 
 // This is called with item._originalUrl when the URL is for a specific data file, ie. dataflow is not used.
@@ -1562,29 +1711,29 @@ function loadAndBuildTable(item, dimensionRequestString) {
             item,
             totalsRequestString
           );
-          promises.push(
-            loadJson(
-              createSdmxJsonResoure(proxyCatalogItemUrl(item, totalsUrl))
-            )
-          );
+          promises.push(item._getData(proxyCatalogItemUrl(item, totalsUrl)));
         }
       }
     }
   } else {
     url = item.originalUrl;
   }
-  promises.push(
-    loadJson(createSdmxJsonResoure(proxyCatalogItemUrl(item, url)))
-  );
+  promises.push(item._getData(proxyCatalogItemUrl(item, url)));
   item.isLoading = true;
   return when
     .all(promises)
     .then(function(jsons) {
       // jsons is an array of length 1 or 2, with optional total data json first, then the specific data json.
       var json = jsons[jsons.length - 1]; // the specific json.
+
+      // The structure of the SDMX 2.1 response is slightly different
+      if (item.sdmxVersionNumber === 2.1) json = json.data;
+
       if (jsons.length === 2) {
         // Process and save the region totals as a datasets object.
         var totalsJson = jsons[0];
+        if (item.sdmxVersionNumber === 2.1) totalsJson = jsons[0].data;
+
         sdmxJsonLib.response.prepare(totalsJson);
         item._regionTotals[totalsRequestString] = {
           dataSets: buildDataSetsFromPreparedJson(item, totalsJson),
@@ -1747,15 +1896,6 @@ function cleanUrl(url) {
 
 function cleanAndProxyUrl(catalogItem, url) {
   return proxyCatalogItemUrl(catalogItem, cleanUrl(url));
-}
-
-function createSdmxJsonResoure(url) {
-  return new Resource({
-    url: url,
-    headers: {
-      Accept: "application/vnd.sdmx.structure+json; version=1.0; charset=utf-8"
-    }
-  });
 }
 
 module.exports = SdmxJsonCatalogItem;
