+++ lib/Models/registerCatalogMembers.js
@@ -32,7 +32,6 @@ var MapboxVectorTileCatalogItem = require("./MapboxVectorTileCatalogItem");
 var OgrCatalogItem = require("./OgrCatalogItem");
 var OpenStreetMapCatalogItem = require("./OpenStreetMapCatalogItem");
 var SdmxJsonCatalogItem = require("./SdmxJsonCatalogItem");
-var Sdmx21JsonCatalogItem = require("./Sdmx21JsonCatalogItem");
 var SensorObservationServiceCatalogItem = require("./SensorObservationServiceCatalogItem");
 var SocrataCatalogGroup = require("./SocrataCatalogGroup");
 var TerriaJsonCatalogFunction = require("./TerriaJsonCatalogFunction");
@@ -103,7 +102,6 @@ var registerCatalogMembers = function() {
     OpenStreetMapCatalogItem
   );
   createCatalogMemberFromType.register("sdmx-json", SdmxJsonCatalogItem);
-  createCatalogMemberFromType.register("sdmx-21-json", Sdmx21JsonCatalogItem);
   createCatalogMemberFromType.register("socrata", SocrataCatalogGroup);
   createCatalogMemberFromType.register(
     "sos",
